refactor : 마인드맵 화면일 때 전용 버튼들 컴포넌트 분리
